GEWA/requirements.txt
torcheval
h5py
PyWavefront
trimesh
open3d
wandb
torch-cluster
pyglet<2
dist/GEWA-1.0-py3-none-any.whl
